requirements.txt
# Core RL dependencies
gym==0.26.0
pygame==2.1.0
numpy>=1.21.0
torch>=1.12.0
torchvision>=0.13.0

# Configuration management
pyyaml>=6.0

# Logging and visualization
tensorboard>=2.10.0
matplotlib>=3.5.0
seaborn>=0.12.0

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Code quality
black>=22.0.0
flake8>=5.0.0
pylint>=2.15.0

# Documentation (optional)
sphinx>=5.0.0
sphinx-rtd-theme>=1.0.0
numpydoc>=1.4.0

# Utilities
tqdm>=4.64.0
pandas>=1.4.0
